fix: update dependencies
the previous PR, #8, updated the generated code, which makes use of new
versions of packages. failing forward and updating the go.mod file to
use the new dependencies.

the changes here are a result of a simple:
```
go mod tidy
```

this upgraded the version of go, presumably because the new versions of
the dependencies require at least go 1.22.7.